requirements.txt
transformers
torch
gradio
pillow
accelerate